изменил int на float в параметре salary

+++ profile.py
@@ -5,7 +5,7 @@ from specialization import Specialization
 class Profile:
     # Это класс в котором будет храниться информация о профиле вакансии
     # или человека ищущего работу
-    def __init__(self, name: str, placer: str, spec: list[Specialization], salary: int, place_work: list,
+    def __init__(self, name: str, placer: str, spec: list[Specialization], salary: float, place_work: list,
                  name_education: str, level_education: int, unique_skills: list[str], stars=None):
         # Это конструктор для класса Profile. name- ФИО, placer - соискатель объявления(компания или человек)
         # spec - специализация это будет лист в виде профессии и ее сужения
